Allow reading logs from the journal instead of /var/log/messages
This introduces a new environment variable: USE_JOURNAL.  If this is
set to "true", fluentd will use the systemd-journal input plugin to
read log messages from the journal instead of /var/log/messages* and
/var/log/containers/*.log.
The immediate benefit is performance - reading log entries directly
from the journal rather than having them go from the journal to
rsyslog to /var/log/messages, or from docker to /var/log/containers/*.log.
Another benefit is the microsecond precision timestamps instead of seconds
used by /var/log/messages.  There is a lot of metadata available from the
journal which we can now use.

This is not yet a complete replacement for the file based input.  There are
still a few issues we need to work out:

* Need to handle log throttling.  Might have to be done at the
  journal level - there aren't any sort of throttling configuration
  options in https://github.com/reevoo/fluent-plugin-systemd
  journald (man journald.conf) has RateLimitInterval=, RateLimitBurst=
  options.  But I don't know if it will be possible to do per-project
  throttling in journald config.
* Need to handle the case if user has configured persistent journal
  logs (man journald.conf - Storage=) so may need to read from
  /var/log/journal in addition to/instead of /run/log/journal
* Need to handle historical logs.  The default configuration for
  journald.conf on el7 is `Storage=auto` which means /run/log/journal
  will be used for `volatile` storage unless the user has created
  /var/log/journal.  `volatile` means not persistent, so the persistent
  logs by default will be /var/log/messages*.  We need some way to grab
  those logs, and de-dup them before storing in elasticsearch.